# agentledger/docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/connector-definitions/custom": {
      "post": {
        "operationId": "ConnectorDefinitionsController_createCustom",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomDefinitionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "ConnectorDefinitions"
        ]
      }
    },
    "/v1/connectors": {
      "get": {
        "operationId": "ConnectorsController_list",
        "parameters": [
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Connectors"
        ]
      },
      "post": {
        "operationId": "ConnectorsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnectorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Connectors"
        ]
      }
    },
    "/v1/connectors/{id}": {
      "get": {
        "operationId": "ConnectorsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Connectors"
        ]
      },
      "patch": {
        "operationId": "ConnectorsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConnectorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Connectors"
        ]
      },
      "delete": {
        "operationId": "ConnectorsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Connectors"
        ]
      }
    },
    "/v1/connectors/{id}/test": {
      "post": {
        "operationId": "ConnectorsController_test",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Connectors"
        ]
      }
    },
    "/v1/connectors/{id}/preview": {
      "post": {
        "operationId": "ConnectorsController_preview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Connectors"
        ]
      }
    },
    "/v1/connectors/{id}/sync": {
      "post": {
        "operationId": "ConnectorsController_sync",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Connectors"
        ]
      }
    },
    "/v1/connectors/{id}/sync-runs": {
      "get": {
        "operationId": "ConnectorsController_syncRuns",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Connectors"
        ]
      }
    },
    "/v1/connectors/{id}/errors": {
      "get": {
        "operationId": "ConnectorsController_errors",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Connectors"
        ]
      }
    },
...
  },
  "info": {
    "title": "AgentLedger Control-Plane API",
    "description": "Tenant-scoped control plane: resources, auth, and analytics over the MVs.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCustomDefinitionDto": {
        "type": "object",
        "properties": {
          "definition": {
            "type": "object"
          }
        },
        "required": [
          "definition"
        ]
      },
      "CreateConnectorDto": {
        "type": "object",
        "properties": {
          "connectorDefinitionId": {
            "type": "string"
          },
          "presetId": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "configJson": {
            "type": "object"
          },
          "mappingOverridesJson": {
            "type": "object"
          },
          "scheduleJson": {
            "type": "object"
          },
          "authSecret": {
            "type": "string"
          },
          "baseUrl": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "displayName"
        ]
      },
      "UpdateConnectorDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "configJson": {
            "type": "object"
          },
          "mappingOverridesJson": {
            "type": "object"
          },
          "scheduleJson": {
            "type": "object"
          },
          "authSecret": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          }
        }
      },
      "TestConnectorDto": {
        "type": "object",
        "properties": {
          "authSecret": {
            "type": "string"
          },
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    }
  ]
}